+++ whatlies/language/_gensim_lang.py
@@ -8,7 +8,7 @@ from gensim.models import KeyedVectors
 
 from whatlies.embedding import Embedding
 from whatlies.embeddingset import EmbeddingSet
-from whatlies.language.common import SklearnTransformerMixin
+from whatlies.language._common import SklearnTransformerMixin
 
 
 class GensimLanguage(SklearnTransformerMixin):
