+++ whatlies/language/_hftransformers_lang.py
@@ -5,7 +5,7 @@ import transformers as trf
 
 from whatlies.embedding import Embedding
 from whatlies.embeddingset import EmbeddingSet
-from whatlies.language.common import SklearnTransformerMixin
+from whatlies.language._common import SklearnTransformerMixin
 
 
 class HFTransformersLanguage(SklearnTransformerMixin):
