+++ whatlies/language/_sense2vec_lang.py
