+++ whatlies/language/_spacy_lang.py
@@ -9,7 +9,7 @@ from sklearn.metrics import pairwise_distances
 
 from whatlies.embedding import Embedding
 from whatlies.embeddingset import EmbeddingSet
-from whatlies.language.common import SklearnTransformerMixin
+from whatlies.language._common import SklearnTransformerMixin
 
 
 class SpacyLanguage(SklearnTransformerMixin):
