+++ whatlies/transformers/_ivis.py
@@ -1,36 +1,44 @@
-import numpy as np
-from sklearn.decomposition import PCA
-
 from whatlies.transformers import Transformer
 from whatlies import EmbeddingSet
-from whatlies.transformers.common import new_embedding_dict
+from whatlies.transformers._common import new_embedding_dict
+
+from ivis import Ivis as IVIS
+import numpy as np
 
 
-class Pca(Transformer):
+class Ivis(Transformer):
     """
     This transformer scales all the vectors in an [EmbeddingSet][whatlies.embeddingset.EmbeddingSet]
-    by means of principal component analysis. We're using the implementation found in
-    [scikit-learn](https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html)
+    by means of Ivis algorithm. We're using the implementation found
+    [here](https://github.com/beringresearch/ivis).
+
+    Important:
+        This language backend might require you to manually install extra dependencies
+        unless you installed via either;
+
+        ```
+        pip install whatlies[ivis]
+        pip install whatlies[all]
+        ```
 
     Arguments:
         n_components: the number of compoments to create/add
-        kwargs: keyword arguments passed to the PCA from scikit-learn
+        kwargs: keyword arguments passed to the [Ivis implementation](https://bering-ivis.readthedocs.io/en/latest/hyperparameters.html)
 
     Usage:
 
     ```python
-    from whatlies.language import SpacyLanguage
-    from whatlies.transformers import Pca
+    from whatlies.language import GensimLanguage
+    from whatlies.transformers import Ivis
 
     words = ["prince", "princess", "nurse", "doctor", "banker", "man", "woman",
              "cousin", "neice", "king", "queen", "dude", "guy", "gal", "fire",
              "dog", "cat", "mouse", "red", "bluee", "green", "yellow", "water",
              "person", "family", "brother", "sister"]
 
-    lang = SpacyLanguage("en_core_web_md")
+    lang = GensimLanguage("wordvectors.kv")
     emb = lang[words]
-
-    emb.transform(Pca(3)).plot_interactive_matrix('pca_0', 'pca_1', 'pca_2')
+    emb.transform(Ivis(3)).plot_interactive_matrix('ivis_0', 'ivis_1', 'ivis_2')
     ```
     """
 
@@ -38,7 +46,8 @@ class Pca(Transformer):
         super().__init__()
         self.n_components = n_components
         self.kwargs = kwargs
-        self.tfm = PCA(n_components=n_components)
+        self.kwargs["verbose"] = 0
+        self.tfm = IVIS(embedding_dims=self.n_components, **self.kwargs)
 
     def fit(self, embset):
         names, X = embset.to_names_X()
@@ -49,7 +58,7 @@ class Pca(Transformer):
     def transform(self, embset):
         names, X = embset.to_names_X()
         new_vecs = self.tfm.transform(X)
-        names_out = names + [f"pca_{i}" for i in range(self.n_components)]
+        names_out = names + [f"ivis_{i}" for i in range(self.n_components)]
         vectors_out = np.concatenate([new_vecs, np.eye(self.n_components)])
         new_dict = new_embedding_dict(names_out, vectors_out, embset)
-        return EmbeddingSet(new_dict, name=f"{embset.name}.pca_{self.n_components}()")
+        return EmbeddingSet(new_dict, name=f"{embset.name}.ivis_{self.n_components}()")
