+++ whatlies/transformers/_normalizer.py
@@ -2,7 +2,7 @@ from sklearn.preprocessing import normalize
 
 from whatlies.transformers import Transformer
 from whatlies import EmbeddingSet
-from whatlies.transformers.common import new_embedding_dict
+from whatlies.transformers._common import new_embedding_dict
 
 
 class Normalizer(Transformer):
