+++ whatlies/transformers/_umap.py
@@ -6,7 +6,7 @@ from numba import NumbaPerformanceWarning
 
 from whatlies.transformers import Transformer
 from whatlies import EmbeddingSet
-from whatlies.transformers.common import new_embedding_dict
+from whatlies.transformers._common import new_embedding_dict
 
 
 class Umap(Transformer):
